requirements.txt
openai>=1.0.0
pandas>=1.3.0
tqdm>=4.60.0
python-dotenv>=0.21.0
supabase>=2.22.0